requirements.txt
amqp==5.1.1
args==0.1.0
async-timeout==4.0.2
billiard==3.6.4.0
bitarray==2.6.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
cryptography==39.0.0
Deprecated==1.2.13
filemagic==1.6
flake8==6.0.0
Flask==2.2.2
Flask-Limiter==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-Testing==0.8.1
greenlet==2.0.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
limits==3.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
ordered-set==4.1.0
packaging==23.0
pdfminer==20191125
pdfminer.six==20221105
pdfreader==0.1.12
pdfrw==0.4
Pillow==9.4.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pyflakes==3.0.1
Pygments==2.14.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
rich==12.6.0
six==1.16.0
SQLAlchemy==1.4.46
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
